feat(utils): refactor utility imports and add toast helper functions
